Fix hitArrays not filling in properly on ship objects
